Envs: Separate phaser render from core environments
