+++ src/envs/classic_control_web/index.ts
@@ -1,0 +1,2 @@
+import { CartPoleEnv } from './cartpole';
+export { CartPoleEnv };
