+++ tsdown.config.ts
@@ -5,6 +5,7 @@ export default defineConfig({
     './src/index.ts',
     './src/spaces/index.ts',
     './src/envs/classic_control/index.ts',
+    './src/envs/classic_control_web/index.ts',
   ],
   minify: {
     compress: true,
